adds the seperate rental listing component
